change view: change field amount from double to Bigdecimal into ExchangeView.java

+++ src/main/java/com/aleks/currency_exchange/view/ExchangeView.java
@@ -10,9 +10,9 @@ public class ExchangeView {
     private Currency targetCurrency;
     private BigDecimal rate;
     private double amount;
-    private double convertedAmount;
+    private BigDecimal convertedAmount;
 
-    public ExchangeView(Currency baseCurrency, Currency targetCurrency, BigDecimal rate, double amount, double convertedAmount) {
+    public ExchangeView(Currency baseCurrency, Currency targetCurrency, BigDecimal rate, double amount, BigDecimal convertedAmount) {
         this.baseCurrency = baseCurrency;
         this.targetCurrency = targetCurrency;
         this.rate = rate;
